+++ src/spline.h
@@ -12,17 +12,17 @@ class Spline: public Interpolador{
 	public:
 		Spline();
 		Spline(int n);
-		Spline(const vector<int> &y);
-		void recalcular(const vector<int> &y);
+		Spline(const vector<double> &y);
+		void recalcular(const vector<double> &y);
 		double evaluar(double x);
 	private:
-		vector<int> a;
+		vector<double> a;
 		vector<double> b;
 		vector<double> c;
 		vector<double> d;
 		vector<vector<double> > factorizacionLU;
 		void generarSistema(int n);
-		void generarSpline(const vector<int> &y);
+		void generarSpline(const vector<double> &y);
 };
 
 #endif // SPLINE_H_INCLUDED
